Removing erroneous code from GetSessionStateItemAsync
There is an expected output parameter, SessionItemLong, that is not being provided. I believe this is a holdover from previous ASPState code that used SessionItemShort and was simply copied over and changed to SessionItemLong. It's no longer needed since SessionItemShort is not used in this implementation and can cause errors if the code is ever hit.
see issue #53 (https://github.com/aspnet/AspNetSessionState/issues/53) for more.